+++ app/AdminModule/presenters/DatasetPresenter.php
@@ -7,7 +7,7 @@ use App\Model\DatasetControl;
 use App\Model\TagsControl;
 use App\Model\UserControl;
 use Nette;
-use App\Components\Forms\BootstrapForm;
+use App\Components\Forms\ProtectedForm;
 use Nette\Application\UI\Form;
 use Nette\Security\Passwords;
 
@@ -29,7 +29,7 @@ final class DatasetPresenter extends BasePresenter
         }
     }
 
-    public function __construct(Nette\Database\Context $db, DatasetControl $datasetControl, CategoryControl $categoryControl, TagsControl $tagsControl, BootstrapForm $BootstrapForm)
+    public function __construct(Nette\Database\Context $db, DatasetControl $datasetControl, CategoryControl $categoryControl, TagsControl $tagsControl, ProtectedForm $BootstrapForm)
     {
         $this->db = $db;
         $this->dataset = $datasetControl;
@@ -40,12 +40,18 @@ final class DatasetPresenter extends BasePresenter
 
     public function renderDefault()
     {
-        $this->template->datasets = $this->dataset->getAll();
+        if ($this->getUser()->isAllowed('global'))
+            $this->template->datasets = $this->dataset->getAll();
+        else
+            $this->template->datasets = $this->dataset->getByUser($this->getUser()->id);
     }
 
     public function actionHide($id)
     {
         try {
+            if (!$this->getUser()->isAllowed('global') && $this->dataset->get($id)->users != $this->getUser()->id)
+                throw new \Exception('Na túto akciu nemáte dostatočné oprávnenie.');
+
             $this->dataset->hide($id);
             $this->flashMessage('Dataset information was updated.', 'success');
         }
@@ -63,35 +69,42 @@ final class DatasetPresenter extends BasePresenter
         $this->template->authors = ['0' => '-- select author --'];
 
         foreach ($authors as $u) {
-            $this->template->authors[$u->id] = $u->name;
+            $this->template->authors[$u->id] = $u->name_sk;
         }
 
         $this->template->categories = [];
         $categories = $this->category->getAll();
 
         foreach ($categories as $c) {
-            $this->template->categories[$c->id] = $c->name;
+            $this->template->categories[$c->id] = $c->name_sk;
         }
 
         $this->template->tags = [];
         $tags = $this->tag->getAll();
 
         foreach ($tags as $t) {
-            $this->template->tags[$t->id] = $t->name;
+            $this->template->tags[$t->id] = $t->name_sk;
         }
     }
 
     protected function createComponentAddForm()
     {
-        $form = $this->BootstrapForm->create();
+        $form = $this->form->create();
 
-        $form->addText('name')->setRequired(true);
+        $form->addText('name_sk')->setRequired(true);
+        $form->addText('name_en')->setRequired(true);
         $form->addText('slug')->setRequired(true);
-        $form->addTextArea('description')->setRequired(true);
+        $form->addTextArea('description_sk')->setRequired(true);
+        $form->addTextArea('description_en')->setRequired(true);
         $form->addSelect('authors', 'Authors', $this->template->authors)->setRequired(true);
         $form->addText('licence')->setRequired(true);
         $form->addSelect('category', 'Category', $this->template->categories)->setRequired(true);
         $form->addMultiSelect('tags', "Tags", $this->template->tags);
+        $form->addText('powerbi');
+        $form->addText('map');
+        $form->addText('year');
+        $form->addText('district');
+        $form->addSelect('onlinedata', 'OnlineData', ['0' => '-- Online Data --', '1' => 'Form', '2' => 'Locations', '3' => 'Summary']);
 
         $form->onSuccess[] = [$this, 'AddFormSucceeded'];
 
@@ -101,7 +114,7 @@ final class DatasetPresenter extends BasePresenter
     public function AddFormSucceeded(Form $form, \stdClass $values)
     {
         try {
-            $this->dataset->create($values->name, $values->slug, $values->description, $values->authors, $values->licence, $values->category, $values->tags, $this->getUser()->id);
+            $this->dataset->create($values->name_sk, $values->name_en, $values->slug, $values->description_sk, $values->description_en, $values->authors, $values->licence, $values->category, $values->tags, $values->powerbi, $values->map, $values->year, $values->district, $values->onlinedata, $this->getUser()->id);
 
             $this->flashMessage('Dataset bol úspešne vytvorený.', 'success');
         }
@@ -125,26 +138,30 @@ final class DatasetPresenter extends BasePresenter
         if (!$this->record)
             $this->redirect('Dataset:');
 
+        if (!$this->getUser()->isAllowed('global') && $this->record->users != $this->getUser()->id) {
+            $this->redirect('Dataset:');
+        }
+
         $authors = $this->db->table('authors')->fetchAll();
 
         $this->template->authors = [];
 
         foreach ($authors as $u) {
-            $this->template->authors[$u->id] = $u->name;
+            $this->template->authors[$u->id] = $u->name_sk;
         }
 
         $this->template->categories = [];
         $categories = $this->category->getAll();
 
         foreach ($categories as $c) {
-            $this->template->categories[$c->id] = $c->name;
+            $this->template->categories[$c->id] = $c->name_sk;
         }
 
         $this->template->tags = [];
         $tags = $this->tag->getAll();
 
         foreach ($tags as $t) {
-            $this->template->tags[$t->id] = $t->name;
+            $this->template->tags[$t->id] = $t->name_sk;
         }
 
         $tags = $this->dataset->getTags($id);
@@ -158,19 +175,28 @@ final class DatasetPresenter extends BasePresenter
         $c['tags'] = $selected_tags;
 
         $this['editForm']->setDefaults($c);
+
+        $this->template->dataset = $this->record;
     }
 
     protected function createComponentEditForm()
     {
-        $form = $this->BootstrapForm->create();
+        $form = $this->form->create();
 
-        $form->addText('name')->setRequired(true);
+        $form->addText('name_sk')->setRequired(true);
+        $form->addText('name_en')->setRequired(true);
         $form->addText('slug')->setRequired(true);
-        $form->addTextArea('description')->setRequired(true);
+        $form->addTextArea('description_sk')->setRequired(true);
+        $form->addTextArea('description_en')->setRequired(true);
         $form->addSelect('authors', 'Authors', $this->template->authors)->setRequired(true);
         $form->addText('licence')->setRequired(true);
+        $form->addText('year');
+        $form->addText('district');
         $form->addSelect('category', 'Category', $this->template->categories)->setRequired(true);
         $form->addMultiSelect('tags', "Tags", $this->template->tags);
+        $form->addText('powerbi');
+        $form->addText('map');
+        $form->addSelect('onlinedata', 'OnlineData', ['0' => '-- Online Data --', '1' => 'Form', '2' => 'Locations', '3' => 'Summary']);
 
         $form->onSuccess[] = [$this, 'EditFormSucceeded'];
 
@@ -180,7 +206,7 @@ final class DatasetPresenter extends BasePresenter
     public function EditFormSucceeded(Form $form, \stdClass $values)
     {
         try {
-            $this->dataset->edit($this->record->id, $values->name, $values->slug, $values->description, $values->authors, $values->licence, $values->category, $values->tags);
+            $this->dataset->edit($this->record->id, $values->name_sk, $values->name_en, $values->slug, $values->description_sk, $values->description_en, $values->authors, $values->licence, $values->category, $values->tags, $values->powerbi, $values->map, $values->year, $values->district, $values->onlinedata);
 
             $this->flashMessage('Dataset bol úspešne upravený.', 'success');
         }
@@ -192,6 +218,9 @@ final class DatasetPresenter extends BasePresenter
     public function actionDelete($id)
     {
         try {
+            if (!$this->getUser()->isAllowed('global') && $this->dataset->get($id)->users != $this->getUser()->id)
+                throw new \Exception('Na túto akciu nemáte dostatočné oprávnenie.');
+
             $this->dataset->delete($id);
             $this->flashMessage('Dataset bol vymazaný.', 'success');
         }
@@ -201,4 +230,32 @@ final class DatasetPresenter extends BasePresenter
 
         $this->redirect('Dataset:');
     }
+
+    /*public function actionMakefiles()
+    {
+        if (!$this->getUser()->isAllowed('admin')) {
+            throw new Nette\Application\ForbiddenRequestException;
+        }
+
+        $ds = $this->db->table('dataset')->fetchAll();
+
+        foreach ($ds as $d) {
+            $files = $this->db->table('dataset_files')->where('dataset', $d->id)->order('created_at')->fetchAll();
+
+            if (!$files)
+                continue;
+
+            $i = 1;
+
+            foreach ($files as $f) {
+                $f->update([
+                    'ord' => $i
+                ]);
+
+                $i++;
+            }
+        }
+
+        die;
+    }*/
 }
